update: posting list model

+++ practice5/src/models/PostingListUnit.py
@@ -2,6 +2,6 @@ class PostingListUnit:
     """
     Define what a posing list cell stored.
     """
-    def __init__(self, document_id:int, frequency:int):
+    def __init__(self, document_id:int, frequency:float):
         self.document_id = document_id
         self.frequency = frequency
